+++ utils/database.py
@@ -19,7 +19,7 @@ else:
     shutil.copyfile("data/loot_template.db", LOOTDB)
     lootDatabase = sqlite3.connect(LOOTDB)
 
-DB_LASTVERSION = 20
+DB_LASTVERSION = 21
 
 
 def init_database():
@@ -244,6 +244,13 @@ def init_database():
             c.execute("ALTER TABLE watched_list ADD reason TEXT")
             c.execute("ALTER TABLE watched_list ADD author INTEGER")
             c.execute("ALTER TABLE watched_list ADD added INTEGER")
+            db_version += 1
+        if db_version == 20:
+            # Joinable ranks
+            c.execute("""CREATE TABLE joinable_roles(
+                server_id INTEGER NOT NULL,
+                role_id INTEGER NOT NULL
+                );""")
             db_version += 1
         print("Updated database to version {0}".format(db_version))
         c.execute("UPDATE db_info SET value = ? WHERE key LIKE 'version'", (db_version,))